+++ modules/3d/misc/java/gen_dict.json
@@ -1,4 +1,7 @@
 {
+    "namespaces_dict": {
+        "cv.fisheye": "fisheye"
+    },
     "func_arg_fix" : {
         "findFundamentalMat"  : { "points1" : {"ctype" : "vector_Point2f"},
                                   "points2" : {"ctype" : "vector_Point2f"} },
