+++ modules/stereo/misc/java/gen_dict.json
@@ -1,0 +1,5 @@
+{
+    "namespaces_dict": {
+        "cv.fisheye": "fisheye"
+    }
+}
